make init script shorter and add exit commends

+++ init_server.sh
@@ -32,7 +32,7 @@ if [ "$EUID" -ne 0 ]; then printf 'Please run as root!\n'; exit 0; fi
 
 # check tun device is available
 if [ ! -e /dev/net/tun ]; then
-	printf '\tThe TUN device is not available!\nYou need to enable TUN before running this script!\n'
+	printf '\tThe TUN device is not available!\nYou need a enabled TUN device (/dev/net/tun) before running this script!\n'
 	exit 0
 fi
 
@@ -40,18 +40,15 @@ fi
 if [ -f /etc/debian_version ]; then
 	PKGMNGR='apt-get';
 elif [ -f /etc/centos-release ] ; then
-	printf 'Centos is not supported yet!\n'; exit 0;
+	printf 'Centos is not supported yet!\n'; exit 0
 	#PKGMNGR='yum'
-elif [ -f /etc/fedora-release ] ; then
-	printf 'Fedora is not supported yet!\n'; exit 0;
-	#PKGMNGR='dnf'
 else
-	printf 'OS not supported yet!\n'; exit 0;
+	printf 'OS is not supported! (only Debian and Ubuntu - see Repository README.md)\n'; exit 0
 fi
 
 
 # update system
-printf '\n### Update System ..\n';
+printf '\n### Update System ..\n'
 
 "$PKGMNGR" update
 "$PKGMNGR" -y upgrade
@@ -59,13 +56,13 @@ printf '\n### Update System ..\n';
 
 
 # install basic software
-printf '\n### Instaĺl Basic Software..\n';
+printf '\n### Instaĺl Basic Software..\n'
 
 "$PKGMNGR" -y install git salt-minion
 
 
 # check users are present
-printf '\n### Check users are present..\n';
+printf '\n### Check users are present..\n'
 
 for users in freifunk syslog
 do
@@ -76,13 +73,11 @@ done
 
 
 # install/update repository
-printf '\n### Install/Update Repository..\n';
+printf '\n### Install/Update Repository..\n'
 
-if [ ! -d "$INSTALL_DIR" ]; then
-	git clone https://github.com/Freifunk-Dresden/ffdd-server "$INSTALL_DIR"
-fi
+test ! -d "$INSTALL_DIR" && git clone https://github.com/Freifunk-Dresden/ffdd-server "$INSTALL_DIR"
+
 cd "$INSTALL_DIR"
-
 	git fetch
 	git checkout "$tag"
 	git pull -f origin "$tag"
@@ -90,24 +85,24 @@ cd "$INSTALL_DIR"
 
 # small helper for salt to create templates (replace: false)
 cp -fv /root/.bashrc /root/.bashrc_bak >/dev/null 2>&1
-mv -v /root/.bash_aliases /root/.bash_aliases_bak >/dev/null 2>&1
-rm -fv /etc/inputrc >/dev/null 2>&1
+test -f /root/.bash_aliases && mv -v /root/.bash_aliases /root/.bash_aliases_bak >/dev/null 2>&1
+mv -v /etc/inputrc /etc/inputrc_bak >/dev/null 2>&1
 
 
 # ensure nvram and nvram.conf are present and correct
-printf '\n### Check "nvram" Setup ..\n';
+printf '\n### Check "nvram" Setup ..\n'
 
 	cp -fv "$INSTALL_DIR"/salt/freifunk/base/nvram/usr/local/bin/nvram /usr/local/bin/
 
 	if [ ! -f /etc/nvram.conf ]; then
-		printf '\n### Create New /etc/nvram.conf and /usr/local/bin/nvram\n';
+		printf '\n### Create New /etc/nvram.conf and /usr/local/bin/nvram\n'
 
 		cp -fv "$INSTALL_DIR"/salt/freifunk/base/nvram/etc/nvram.conf /etc/nvram.conf
 
 	else
 		# Temp.-Part to update old servers
-		printf '\n### /etc/nvram.conf exists.\n';
-		printf '### Create /etc/nvram.conf.default & /etc/nvram.conf.diff\n';
+		printf '\n### /etc/nvram.conf exists.\n'
+		printf '### Create /etc/nvram.conf.default & /etc/nvram.conf.diff\n'
 
 		# check new options are set
 		# check autoupdate
@@ -132,18 +127,14 @@ printf '\n### Check "nvram" Setup ..\n';
 
 	# check default Interface is correct set
 	get_default_interface
-	if [ "$(nvram get ifname)" != "$def_if" ]; then
-		nvram set ifname "$def_if"
-	fi
+	[ "$(nvram get ifname)" != "$def_if" ] && nvram set ifname "$def_if"
 
 	# check install_dir is correct set
-	if [ "$(nvram get install_dir)" != "$INSTALL_DIR" ]; then
-		nvram set install_dir "$INSTALL_DIR"
-	fi
+	[ "$(nvram get install_dir)" != "$INSTALL_DIR" ] && nvram set install_dir "$INSTALL_DIR"
 
 
 # create clean masterless salt enviroment
-printf '\n### Check Salt Enviroment ..\n';
+printf '\n### Check Salt Enviroment ..\n'
 
 rm -fv /etc/salt/minion.d/*.conf
 
@@ -162,7 +153,7 @@ EOF
 
 
 # ensure running services are stopped
-printf '\n### Ensure Services are Stopped ..\n';
+printf '\n### Ensure Services are Stopped ..\n'
 
 services='S40network S41firewall S42firewall6 S52batmand S53backbone-fastd2 S90iperf3 fail2ban apache2 monitorix openvpn@openvpn-vpn0 openvpn@openvpn-vpn1 wg-quick@vpn0 wg-quick@vpn1'
 for s in $services
@@ -181,7 +172,7 @@ done
 # -- Initial System --
 #
 
-printf '\n### Start Initial System. (please wait! Coffee Time ca.10min)\n';
+printf '\n### Start Initial System. (please wait! Coffee Time ca.10min)\n'
 
 salt-call state.highstate --local
 
@@ -190,7 +181,7 @@ salt-call state.highstate --local
 # -- Cleanup System --
 #
 
-printf '\n### .. All done! Cleanup System ..\n';
+printf '\n### .. All done! Cleanup System ..\n'
 
 "$PKGMNGR" autoremove
